Make Capybara::UI a module and use Capybara::UI::Page for class
